optimize(bytesconv): add support for go1.20
